Create code
This project focuses on predicting loan approvals using various machine learning algorithms to analyze and classify data from a dataset. The dataset, LoanApprovalPrediction.csv, contains information about applicants, including categorical variables such as gender, marital status, and loan status, as well as numerical features like loan amount and income. The objective of this project is to build a predictive model that can classify whether a loan application will be approved or not.